+++ lib/secp256k1/secp256k1.ex
@@ -113,7 +113,17 @@ defmodule Bitcoinex.Secp256k1 do
 
     @spec serialize_signature(t()) :: binary
     def serialize_signature(%__MODULE__{r: r, s: s}) do
-      :binary.encode_unsigned(r) <> :binary.encode_unsigned(s)
+      Utils.int_to_big(r, 32) <> Utils.int_to_big(s, 32)
+    end
+
+    @doc """
+    to_hex returns the compact 64-byte (r || s) signature as a lowercase hex string.
+    """
+    @spec to_hex(t()) :: String.t()
+    def to_hex(sig = %__MODULE__{}) do
+      sig
+      |> serialize_signature()
+      |> Base.encode16(case: :lower)
     end
 
     @doc """
@@ -205,7 +215,7 @@ defmodule Bitcoinex.Secp256k1 do
           if Point.has_even_y(pubkey) do
             privkey
           else
-            %PrivateKey{d: Params.curve().n - privkey.d}
+            PrivateKey.negate(privkey)
           end
         end
     end